fix: changed "ft_putnbr_base" number type to admint pointers length

+++ ft_printf/ft_putnbr_base.c
@@ -6,15 +6,16 @@
 /*   By: judelgad <judelgad@student.42malaga.com>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/23 17:45:19 by judelgad          #+#    #+#             */
-/*   Updated: 2023/12/26 10:45:18 by judelgad         ###   ########.fr       */
+/*   Updated: 2023/12/26 17:23:05 by judelgad         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 #include "libft/libft.h"
 #include <unistd.h>
+#include <stdint.h>
 
-int	ft_putnbr_base(unsigned int n, const char *base, int *count)
+int	ft_putnbr_base(uintptr_t n, const char *base, int *count)
 {
 	size_t	base_len;
